+++ laravel-domains/src/LaravelDomainsServiceProvider.php
@@ -3,58 +3,41 @@
 namespace Pushberryfinn\LaravelDomains;
 
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
 
 class LaravelDomainsServiceProvider extends ServiceProvider
 {
-    /**
-     * Bootstrap the application services.
-     */
+
     public function boot()
     {
-        /*
-         * Optional methods to load your package assets
-         */
-        // $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'laravel-domains');
-        // $this->loadViewsFrom(__DIR__.'/../resources/views', 'laravel-domains');
-        // $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
-        // $this->loadRoutesFrom(__DIR__.'/routes.php');
-
         if ($this->app->runningInConsole()) {
-            $this->publishes([
-                __DIR__.'/../config/config.php' => config_path('laravel-domains.php'),
-            ], 'config');
-
-            // Publishing the views.
-            /*$this->publishes([
-                __DIR__.'/../resources/views' => resource_path('views/vendor/laravel-domains'),
-            ], 'views');*/
-
-            // Publishing assets.
-            /*$this->publishes([
-                __DIR__.'/../resources/assets' => public_path('vendor/laravel-domains'),
-            ], 'assets');*/
-
-            // Publishing the translation files.
-            /*$this->publishes([
-                __DIR__.'/../resources/lang' => resource_path('lang/vendor/laravel-domains'),
-            ], 'lang');*/
-
-            // Registering package commands.
-            // $this->commands([]);
+            $this->registerConsoleCommands();
         }
+
     }
 
-    /**
-     * Register the application services.
-     */
-    public function register()
+    protected function registerConsoleCommands()
     {
-        // Automatically apply the package configuration
-        $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'laravel-domains');
+        $commandNamespace = 'Pushberryfinn\\LaravelDomains\\Console\\';
+        $commandPath = __DIR__ . '/Console';
 
-        // Register the main class to use with the facade
-        $this->app->singleton('laravel-domains', function () {
-            return new LaravelDomains;
-        });
+        if (!is_dir($commandPath)) {
+            return;
+        }
+
+        $commands = [];
+
+        foreach (File::files($commandPath) as $file) {
+            $class = $commandNamespace . Str::replaceLast('.php', '', $file->getFilename());
+
+            if (class_exists($class) && is_subclass_of($class, \Illuminate\Console\Command::class)) {
+                $commands[] = $class;
+            }
+        }
+
+        if (!empty($commands)) {
+            $this->commands($commands);
+        }
     }
 }